dom/base/test/jsmodules/test_scriptNotParsedAsModule.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test classic script not parsed as module</title>
<script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var wasRun = false;
  var hadSyntaxError = false;

  SimpleTest.waitForExplicitFinish();
  window.onerror = handleError;

  function handleError(message, url, line, column, error) {
    hadSyntaxError = error instanceof SyntaxError;
  }

  function testError() {
    ok(!wasRun, 'Check script was not run');
    ok(hadSyntaxError, 'Check that a SyntaxError was thrown');
    SimpleTest.finish();
  }
</script>
<script src="module_testSyntax.js"></script>
<body onload='testError()'></body>
